+++ .gitmodules
@@ -1,0 +1,3 @@
+[submodule "lib/wolfssl"]
+	path = lib/wolfssl
+	url = https://github.com/wolfssl/wolfssl
