+++ tox.ini
@@ -1,6 +1,6 @@
 [tox]
-envlist = py27, py34, py35, py36
-skip_missing_interpreters = true
+envlist = py39, pep8
+skipsdist = True
 
 [testenv]
 setenv =
